Add command-line arguments
The network interface to listen on may not be always the same
(in my case, was wlan1).
The queue service may not necessarily be hosted on localhost,
or one may not want to use it at all, temporarily.

Now, I'm not sure the added complexity is worth it, when users
of this project are likely to customize and recompile it.